refactor project structure to MVC
