+++ src/main.py
@@ -4,7 +4,7 @@ from fastapi import FastAPI
 from fastapi.middleware.cors import CORSMiddleware
 from fastapi.staticfiles import StaticFiles
 
-from src.api.endpoints import router
+from src.controls.endpoints import router
 
 app = FastAPI(title="Ebla ChatBot Assistant")
 app.include_router(router)
@@ -12,7 +12,7 @@ app.include_router(router)
 BASE_DIR = Path(__file__).parent
 
 # Mount static files (CSS, images)
-app.mount("/static", StaticFiles(directory=BASE_DIR / "static"), name="static")
+app.mount("/static", StaticFiles(directory=BASE_DIR / "view/static"), name="static")
 
 # Allow CORS for frontend applications
 app.add_middleware(
