+++ src/view/static/style.css
